+++ src/OpenTelemetry.AutoInstrumentation/Instrumentation.cs
@@ -124,7 +124,7 @@ internal static class Instrumentation
                 }
                 else
                 {
-                    AddLazilyLoadedTraceInstrumentations(LazyInstrumentationLoader, _pluginManager, TracerSettings.Value.EnabledInstrumentations);
+                    AddLazilyLoadedTraceInstrumentations(LazyInstrumentationLoader, _pluginManager, TracerSettings.Value);
                     Logger.Information("Initialized lazily-loaded trace instrumentations without initializing sdk.");
                 }
             }
@@ -256,9 +256,9 @@ internal static class Instrumentation
         }
     }
 
-    private static void AddLazilyLoadedTraceInstrumentations(LazyInstrumentationLoader lazyInstrumentationLoader, PluginManager pluginManager, IReadOnlyList<TracerInstrumentation> enabledInstrumentations)
+    private static void AddLazilyLoadedTraceInstrumentations(LazyInstrumentationLoader lazyInstrumentationLoader, PluginManager pluginManager, TracerSettings tracerSettings)
     {
-        foreach (var instrumentation in enabledInstrumentations)
+        foreach (var instrumentation in tracerSettings.EnabledInstrumentations)
         {
             switch (instrumentation)
             {
@@ -296,10 +296,11 @@ internal static class Instrumentation
                     break;
                 case TracerInstrumentation.MassTransit:
                     break;
+                case TracerInstrumentation.GraphQL:
+                    DelayedInitialization.Traces.AddGraphQL(LazyInstrumentationLoader, pluginManager, tracerSettings);
+                    break;
 #endif
                 case TracerInstrumentation.MongoDB:
-                    break;
-                case TracerInstrumentation.GraphQL:
                     break;
                 case TracerInstrumentation.Npgsql:
                     break;