+++ test/test-applications/integrations/TestApplication.GraphQL/StarWars/Types/EpisodeEnum.cs
@@ -1,0 +1,22 @@
+using GraphQL.Types;
+
+namespace StarWars.Types;
+
+public enum Episodes
+{
+    NEWHOPE = 4,
+    EMPIRE = 5,
+    JEDI = 6
+}
+
+public class EpisodeEnum : EnumerationGraphType
+{
+    public EpisodeEnum()
+    {
+        Name = "Episode";
+        Description = "One of the films in the Star Wars Trilogy.";
+        Add("NEWHOPE", 4, "Released in 1977.");
+        Add("EMPIRE", 5, "Released in 1980.");
+        Add("JEDI", 6, "Released in 1983.");
+    }
+}